Moana gateway working
