+++ lib/index.js
@@ -66,10 +66,9 @@ async function generalRequest(url, method, body, fullResponse) {
  * @param {Array<string>} mutations
  * @return {string}
  */
-function mergeSchemas(typeDefs, queries, mutations) {
+function mergeSchemas(typeDefs, queries) {
 	return `${typeDefs.join('\n')}
-    type Query { ${queries.join('\n')} }
-    type Mutation { ${mutations.join('\n')} }`;
+    type Query { ${queries.join('\n')} }`;
 }
 
 function formatErr(error) {
@@ -89,43 +88,25 @@ type Post {
     createdAt: String!
     idCreator: String!
     content: String!
-}
-input PostUpdate {
-    newContent: String!
-}
-input PostInput {
-    idCreator: String!
-    content: String!
 }`;
 
-const postsQueries = `
-    postById(id: String!): Post!
-`;
-
-const postsMutations = `
-    createPost(post: PostInput!): Post!
-    deletePost(id: String!): Post!
-    updatePost(id: String!, newContent: PostUpdate!): Post!
+const feedQueries = `
+    homeFeedForUser(id: String!): [Post]!
+    profileFeedForUser(id: String!): [Post]!
 `;
 
 const url = 'localhost';
 const port = '3000';
-const entryPoint = 'posts';
+const entryPoint = 'feeds';
 
 const URL = `http://${url}:${port}/${entryPoint}`;
 
 const resolvers = {
 	Query: {
-		postById: (_, { id }) =>
-			generalRequest(`${URL}/find/${id}`, 'GET'),
-	},
-	Mutation: {
-		createPost: (_, { post }) =>
-			generalRequest(`${URL}`, 'POST', post),
-		updatePost: (_, { id, newContent }) =>
-			generalRequest(`${URL}/update/${id}`, 'PUT', newContent),
-		deletePost: (_, { id }) =>
-			generalRequest(`${URL}/delete/${id}`, 'DELETE')
+		homeFeedForUser: (_, { id }) =>
+			generalRequest(`${URL}/home/${id}`, 'GET'),
+		profileFeedForUser: (_, { id }) =>
+			generalRequest(`${URL}/profile/${id}`, 'GET')
 	}
 };
 
@@ -136,10 +117,7 @@ const mergedTypeDefs = mergeSchemas(
 		postsTypeDef
 	],
 	[
-		postsQueries
-	],
-	[
-		postsMutations
+		feedQueries
 	]
 );
