+++ src/moana/server.js
@@ -1,3 +1,3 @@
 export const url = 'localhost'
 export const port = '3000'
-export const entryPoint = 'posts'
+export const entryPoint = 'feeds'
