+++ src/moana/typeDefs.js
@@ -1,0 +1,12 @@
+export const postsTypeDef = `
+type Post {
+    id: String!
+    createdAt: String!
+    idCreator: String!
+    content: String!
+}`;
+
+export const feedQueries = `
+    homeFeedForUser(id: String!): [Post]!
+    profileFeedForUser(id: String!): [Post]!
+`;
