+++ src/utilities.js
@@ -72,10 +72,9 @@ export function getRequest(url, path, parameters) {
  * @param {Array<string>} mutations
  * @return {string}
  */
-export function mergeSchemas(typeDefs, queries, mutations) {
+export function mergeSchemas(typeDefs, queries) {
 	return `${typeDefs.join('\n')}
-    type Query { ${queries.join('\n')} }
-    type Mutation { ${mutations.join('\n')} }`;
+    type Query { ${queries.join('\n')} }`;
 }
 
 export function formatErr(error) {
